# Conflicts:
#	ERP_V2/src/main/java/com/example/ERP_V2/DTO/OrderDTO.java

+++ ERP_V2/src/main/java/com/example/ERP_V2/Model/Order.java
@@ -87,7 +87,7 @@ public class Order {
     private BigDecimal outerPaperRate;
 
     @Column(precision = 6, scale = 2)
-    private BigDecimal estimatedAmount;
+    private Long estimatedAmount;
 
     @ManyToOne
     @JoinColumn(name = "customer_id")
